requirements.txt
Flask
Werkzeug
gunicorn

opencv-python
numpy>=1.26.0

Pillow

torch
torchvision

paddleocr
paddlepaddle

spacy

fuzzywuzzy
python-Levenshtein

scikit-image

fpdf

transformers
sentencepiece
sacremoses
